api/requirements.txt
fastapi
strawberry-graphql[fastapi]
sqlalchemy
asyncpg
alembic
uvicorn
python-dotenv
psycopg
ruff
